updates `accounts.user` to be `users.user` integrated with pow
